==> 01_VG_companies_distribution.ipynb <==
# Videogame companies distribution in USA

Since most of the videogame industry is placed in the United States, I'd like to figure out where are they located.

from pymongo import MongoClient
client = MongoClient("mongodb://localhost:27017/ironhack")
db = client.get_database()

from collections import Counter

query = {"category_code": "games_video", "offices.country_code":"USA"}
project = {"_id":0,"offices.city":1}
result = db["companies"].find(query,project)
usa_vg= list (result)
print("There are ",len(usa_vg), " companies","    ", type(usa_vg))


cities = []
for i in usa_vg:
    for k,v in i.items():
        for c in v:
            cities.append(c["city"])
Counter(cities)

The city with more video_game companies in USA is New York, but it is the only city in the east coast that has wide developed this kind of industry while in California, although the companies are distributed in diferent cities there are a lot more of them nearer.
Taking into account all this, we are going to put the company in San Francisco.


That´s why I'll choose San Francisco to place my company
