==> 02_SF_videogame_companies.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Mongo-request" data-toc-modified-id="Mongo-request-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Mongo request</a></span></li><li><span><a href="#Dataframe-cleaning" data-toc-modified-id="Dataframe-cleaning-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Dataframe cleaning</a></span></li><li><span><a href="#Map-visualization" data-toc-modified-id="Map-visualization-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Map visualization</a></span></li></ul></div>

# San Francisco videogame companies

Most of our teams are developers and designers, we want to make them comfortable so it is important to consider their preferences.
That's why we are going to place the office in a zone where there are other emergent videogame companies that have rised at least 1 milliondollars. Moreover, focusing on videogame companies instead of any tech companies, we ensure that they also have designing interests so if they organize talks on the subject, it would be easy for our designers to attend. 

We are going to use Mongodb "companies".


from pymongo import MongoClient
client = MongoClient("mongodb://localhost:27017/ironhack")
db = client.get_database()

import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster

companies = client.companies.companies
offices = client.companies.offices

import pandas as pd

## Mongo request

query = {"category_code": "games_video"}
project = {"_id":0, "name":1,"offices.city":1, "founded_year":1}
result = db["companies"].find(query,project).sort("founded_year",-1)
x= list (result)
len(x)
x[:4]

On a first overview we can deduce that the newest video games companies have been founded before 2011 so we can´t consider them exactly startups but let's consider those that are relatively new, for example those founded since 2005.
We will take into account those that have raise at least 1 million dollars, so the tech developers are happy.

query = {"category_code": "games_video",  "founded_year": {"$gte" : 2005}, "offices.city":"San Francisco", "total_money_raised":{"$regex" : "M"}}
project = {"_id":0, "name":1,"offices":1, "total_money_raised":1, "founded_year":1}
result = db["companies"].find(query,project)
sf_videogames= list (result)
len(sf_videogames)

sf_videogames[0]

sf_df = pd.json_normalize(sf_videogames,"offices",["name", "founded_year", "total_money_raised"])
sf_df.head()

## Dataframe cleaning

Although we have already filtered the companies depending on if they had an office in San Francisco or not, those that have more than one offfice are apparing in our dataframe, so we need to drop them.

# Get indexes where name column doesn't have value 'San Francisco'
indexNames = sf_df[~(sf_df['city'] == 'San Francisco')].index 
# Delete these row indexes from dataFrame
sf_df.drop(indexNames , inplace=True)
sf_df = sf_df[sf_df['latitude'].notna()]
sf_df = sf_df[sf_df['longitude'].notna()]
sf_df.head()

len(sf_df)

import geopandas as gpd

sf_gdf = gpd.GeoDataFrame(sf_df,geometry = gpd.points_from_xy(sf_df.longitude, sf_df.latitude))

sf_gdf.head()

sf_gdf.to_csv("data/san_francisco_vg_comp_gdf.csv")

## Map visualization

inicial_lat = 37.7825098
inicial_long = -122.4077973

map_1 = folium.Map(location = [inicial_lat,inicial_long], zoom_start = 15)


HeatMap(data = sf_df [["latitude", "longitude"]], radius =15).add_to(map_1)
map_1

We can see that these companies tend to accumulate in three areas. Let's find out if we can satify more requirements near here.

#Van Ness avenue with Market Street
loc1 = 37.77501,-122.4167646
#2nd with Bryan Street more or les
loc2 =37.7832303,-122.3936295
#Motgomery St. Station
loc3 =37.7887223,-122.4009011
