==> 03_Foursquare_API_calls.ipynb <==
# Collecting data from the Foursquare API

<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Looking-for-Starbucks" data-toc-modified-id="Looking-for-Starbucks-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Looking for Starbucks</a></span><ul class="toc-item"><li><span><a href="#Extract-geopandas-dataframe" data-toc-modified-id="Extract-geopandas-dataframe-1.1"><span class="toc-item-num">1.1&nbsp;&nbsp;</span>Extract geopandas dataframe</a></span></li><li><span><a href="#Convert-geodataframe-in-geojson-format" data-toc-modified-id="Convert-geodataframe-in-geojson-format-1.2"><span class="toc-item-num">1.2&nbsp;&nbsp;</span>Convert geodataframe in geojson format</a></span></li><li><span><a href="#Discover-which-is-the-nearest-item-to-my-initial-places" data-toc-modified-id="Discover-which-is-the-nearest-item-to-my-initial-places-1.3"><span class="toc-item-num">1.3&nbsp;&nbsp;</span>Discover which is the nearest item to my initial places</a></span></li></ul></li><li><span><a href="#Looking-for-airports" data-toc-modified-id="Looking-for-airports-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Looking for airports</a></span><ul class="toc-item"><li><span><a href="#Extract-geopandas-dataframe" data-toc-modified-id="Extract-geopandas-dataframe-2.1"><span class="toc-item-num">2.1&nbsp;&nbsp;</span>Extract geopandas dataframe</a></span></li><li><span><a href="#Convert-geodataframe-in-geojson-format" data-toc-modified-id="Convert-geodataframe-in-geojson-format-2.2"><span class="toc-item-num">2.2&nbsp;&nbsp;</span>Convert geodataframe in geojson format</a></span></li></ul></li><li><span><a href="#Looking-for-veterinarian-services-for-Dobby" data-toc-modified-id="Looking-for-veterinarian-services-for-Dobby-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Looking for veterinarian services for Dobby</a></span></li><li><span><a href="#Looking-for-elementary-schools" data-toc-modified-id="Looking-for-elementary-schools-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Looking for elementary schools</a></span></li><li><span><a href="#Looking-for-vegan-restaurant-for-the-CEO" data-toc-modified-id="Looking-for-vegan-restaurant-for-the-CEO-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Looking for vegan restaurant for the CEO</a></span></li><li><span><a href="#Looking-for-after-COVID-possible-party-places" data-toc-modified-id="Looking-for-after-COVID-possible-party-places-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Looking for after-COVID possible party places</a></span></li><li><span><a href="#Looking-for-a-basketball-stadium-for-the-maintenance-guy" data-toc-modified-id="Looking-for-a-basketball-stadium-for-the-maintenance-guy-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>Looking for a basketball stadium for the maintenance guy</a></span></li></ul></div>

Let's get some information from Forusquare API about all the givem requirements.

This requirements are:

- Designers like to go to design talks and share knowledge. There must be some nearby companies that also do design.
- 30% of the company staff have at least 1 child.
- Developers like to be near successful tech startups that have raised at least 1 Million dollars.
- Executives like Starbucks A LOT. Ensure there's a starbucks not too far.
- Account managers need to travel a lot.
- Everyone in the company is between 25 and 40, give them some place to go party.
- The CEO is vegan.
- If you want to make the maintenance guy happy, a basketball stadium must be around 10 Km.
- The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away.

For that i'm using the "api_to_geopd" function defined in src folder on the file called foursquare.py .
This function makes a call to the API and trasforms  de obteined data into a goepandas dataframe.

After that we are going to transform each dataframe into a mongo collection because after that we want to make some geoqueries.

In previous files we have researched which locations could be the best option for our company. We reached to this three options in **San Francisco, CA, USA**:

 **Van Ness avenue with Market Street**
 
    loc1 = 37.77501,-122.4167646
 **2nd with Bryan Street more or les**
 
    loc2 =37.7832303,-122.3936295
 **Motgomery St. Station**
 
    loc3 =37.7887223,-122.4009011
    
    
The three potential locations are located in San Francisco, if I don't specify th "radius", Foursquare API gives by default all the items limited to the city, so I can use the information of each call for all the locations.

import sys
sys.path.append("../")
from src.foursquare import *

from dotenv import load_dotenv
import os

import shapely.geometry

from pymongo import MongoClient,GEOSPHERE
client = MongoClient("mongodb://localhost:27017/ironhack")

*---In order to keep safe my Foursquare tokens, i store them in a separated file that it's not going to be uploaded---*

load_dotenv()

tok1 = os.getenv("tok1")
tok2 = os.getenv("tok2")

## Looking for Starbucks 

### Extract geopandas dataframe

starbucks= api_to_geopd ("Starbucks",50,tok1,tok2)

starbucks[:3]

### Convert geodataframe in geojson format

gpd_to_geojson(starbucks, "starbucks")
starbucks.head()

### Discover which is the nearest item to my initial places

loc1 = [-122.4167646,37.77501]

db=client.ironhack

nearest = db.starbucks.find({"geometry":{"$near":{"type":"Point", "coordinates":[-122.4167646,37.77501]}}})
list(nearest)[0]

## Looking for airports 

### Extract geopandas dataframe

airports = api_to_geopd ("airport",1,tok1,tok2)
airports[:3]

### Convert geodataframe in geojson format

gpd_to_geojson(airports, "airports")
airports.head()

## Looking for veterinarian services for Dobby 

veterinary = api_to_geopd ("veterinarian",50,tok1,tok2)
veterinary [:3]

gpd_to_geojson(veterinary, "vet")
veterinary.head()

## Looking for elementary schools
Keeping in mind that 30% of the company staff have at least 1 child, and that they are between 25 and 40 I supose that the childs are in elementary school more or less.

elementary_school = api_to_geopd ("elementary school",50,tok1,tok2)
elementary_school[:3]

gpd_to_geojson(elementary_school, "schools")
elementary_school.head()

## Looking for vegan restaurant for the CEO

vegan = api_to_geopd ("vegan restaurant",50,tok1,tok2)
vegan[:3]

gpd_to_geojson(vegan, "vegan")
vegan.head()

## Looking for after-COVID possible party places

nightclub = api_to_geopd ("nightclub",50,tok1,tok2)
nightclub[:3]

gpd_to_geojson(nightclub, "party")
nightclub.head()

## Looking for a basketball stadium for the maintenance guy

basketball = api_to_geopd ("basketball stadium",5,tok1,tok2)
basketball

gpd_to_geojson(basketball, "basket_stadiums")
basketball.head()
