==> 04_nearest_services.ipynb <==
# Nearest services

<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Looking-for-the-nearest-services-in-&quot;Van-Ness-avenue-with-Market-Street&quot;" data-toc-modified-id="Looking-for-the-nearest-services-in-&quot;Van-Ness-avenue-with-Market-Street&quot;-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Looking for the nearest services in "Van Ness avenue with Market Street"</a></span></li><li><span><a href="#Looking-for-the-nearest-services-in-&quot;2nd-with-Bryan-Street&quot;" data-toc-modified-id="Looking-for-the-nearest-services-in-&quot;2nd-with-Bryan-Street&quot;-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Looking for the nearest services in "2nd with Bryan Street"</a></span></li><li><span><a href="#Looking-for-the-nearest-services-near-&quot;Montgomery-St.-Station&quot;" data-toc-modified-id="Looking-for-the-nearest-services-near-&quot;Montgomery-St.-Station&quot;-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Looking for the nearest services near "Montgomery St. Station"</a></span></li></ul></div>

In this notebook I'm using all the data collected in the previous notebooks to agreggate the nearest servicies required to each of the three potential locations.

For this I'm using de GeoJSON collections I created in the previous notebook.

import sys
sys.path.append("../")
from src.foursquare import *
from pymongo import MongoClient,GEOSPHERE
client = MongoClient("mongodb://localhost:27017/ironhack")
import pandas as pd

 **Van Ness avenue with Market Street**
 
    loc1 = -122.4167646, 37.77501
 **2nd with Bryan Street more or les**
 
    loc2 = -122.3936295, 37.7832303
 **Motgomery St. Station**
 
    loc3 =-122.4009011, 37.7887223

db=client.ironhack

## Looking for the nearest services in "Van Ness avenue with Market Street"

loc1 = [-122.4167646,37.77501]
loc1_list = []

# Nearest Starbucks
nearest = db.starbucks.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest Vet for Dobby
nearest = db.vet.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest schools
nearest = db.schools.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest vegan restaurants
nearest = db.vegan.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest nightclubs
nearest = db.party.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

# Nearest basketball stadium
nearest = db.basket_stadiums# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc1}}})
x = list(nearest)[0]
print(x)
loc1_list.append(x)

Now I make a dataframe with all the services

loc1_df = pd.DataFrame(loc1_list)
loc1_df

loc1_df.to_csv("data/loc1_nearest_requirements.csv")

## Looking for the nearest services in "2nd with Bryan Street"

loc2 = [-122.3936295, 37.7832303]
loc2_list = []

# Nearest Starbucks
nearest = db.starbucks.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest Vet for Dobby
nearest = db.vet.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest schools
nearest = db.schools.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest vegan restaurants
nearest = db.vegan.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest nightclubs
nearest = db.party.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

# Nearest basketball stadium
nearest = db.basket_stadiums# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc2}}})
a = list(nearest)[0]
print(a)
loc2_list.append(a)

loc2_df = pd.DataFrame(loc2_list)
loc2_df

loc2_df.to_csv("data/loc2_nearest_requirements.csv")

## Looking for the nearest services near "Montgomery St. Station"

loc3 = [-122.4009011, 37.7887223]
loc3_list = []

# Nearest Starbucks
nearest = db.starbucks.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest Vet for Dobby
nearest = db.vet.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest schools
nearest = db.schools.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest vegan restaurants
nearest = db.vegan.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest nightclubs
nearest = db.party.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

# Nearest basketball stadium
nearest = db.basket_stadiums# Nearest airport
nearest = db.airports.find({"geometry":{"$near":{"type":"Point", "coordinates":loc3}}})
x = list(nearest)[0]
print(x)
loc3_list.append(x)

loc3_df = pd.DataFrame(loc3_list)
loc3_df

loc3_df.to_csv("data/loc3_nearest_requirements.csv")

'''**distances**
import json
from bson.json_util import dumps
query =[{'$geonear':{'near':loc1,
                    'distanceField': "distance",
                    'maxDistance':15000,
                    'distanceMultiplier': 6371,
                    'spherical':True}}]
nearest2 = db.starbucks.aggregate({"geometry":{'$geonear':{'near':loc1,
                                                           'distanceField': "distance",
                                                           'maxDistance':15000,
                                                           'distanceMultiplier': 6371,
                                                           'spherical':True}}})

response_json = json.loads(dumps(geoloc))'''
