==> 05_Map_visualization.ipynb <==
# Map visualization

This notebook will reflect on a map all data obtained in previous notebooks

import sys
sys.path.append("../")
import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster
import pandas as pd

## First location: Van Ness avenue with Market Street

loc1 = -122.4167646, 37.77501

loc1_df = pd.read_csv('data/loc1_nearest_requirements.csv')
loc1_df

## Second location: 2nd with Bryan Street

loc2 = -122.3936295, 37.7832303

loc2_df = pd.read_csv('data/loc2_nearest_requirements.csv')
loc2_df

## Third location: Montgomery St. Station

loc3 =-122.4009011, 37.7887223

loc3_df = pd.read_csv('data/loc3_nearest_requirements.csv')
loc3_df

## Let's paint it on a map

First, the three potential locations for our office will be added to the map.

inicial_lat = 37.7825098
inicial_long = -122.4077973
map_1 = folium.Map(location = [inicial_lat,inicial_long], zoom_start = 13)

loc_one = Marker (location= [loc1[1],loc1[0]], tooltip="Van Ness avenue with Market Street")
loc_one.add_to(map_1)

loc_two = Marker (location= [loc2[1],loc2[0]], tooltip="2nd with Bryan Street")
loc_two.add_to(map_1)

loc_three = Marker (location= [loc3[1],loc3[0]], tooltip="Montgomery St. Station")
loc_three.add_to(map_1)

Now, we will paint all the other locations on the map.

for i,row in loc1_df.iterrows():
    pubs = {
        "location" : [row["latitud"],row["longitud"]],
        "tooltip" : row["name"]
    }
    icon = Icon(color = "black",
                    prefix = "fa",
                    icon = "bullseye",
                    icon_color = "white"
        )
    Marker(**pubs,icon=icon).add_to(map_1)
    
    
    

for i,row in loc2_df.iterrows():
    pubs = {
        "location" : [row["latitud"],row["longitud"]],
        "tooltip" : row["name"]
    }
    icon = Icon(color = "red",
                    prefix = "fa",
                    icon = "bullseye",
                    icon_color = "white"
        )
    Marker(**pubs,icon=icon).add_to(map_1)
    
    
    
    
for i,row in loc3_df.iterrows():
    pubs = {
        "location" : [row["latitud"],row["longitud"]],
        "tooltip" : row["name"]
    }
    icon = Icon(color = "green",
                    prefix = "fa",
                    icon = "bullseye",
                    icon_color = "white"
        )
    Marker(**pubs,icon=icon).add_to(map_1)
   
map_1

On the map, the potential locations of the office are marked in blue.

For each location, the closest required services have been indicated in a different color.

It should be noted that the closest veterinarian coincides in the first and third locations, so the markers corresponding to the closest green and black veterinarian overlap.

After inspecting the previous map we can deduce that the location that best satisfies all the requirements is Van Ness avenue with Market Street.
